# rimway-backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rides/estimate": {
      "post": {
        "operationId": "RideController_estimateFare",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateFareDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ride"
        ]
      }
    },
    "/api/v1/rides/request": {
      "post": {
        "operationId": "RideController_requestRide",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ride"
        ]
      }
    },
    "/api/v1/rides/cancel": {
      "post": {
        "operationId": "RideController_cancelRide",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ride"
        ]
      }
    },
...
    "/api/v1/rides/{id}/rate": {
      "post": {
        "operationId": "RideController_rateRide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ride"
        ]
      }
    },
...
    "/api/v1/drivers/register": {
      "post": {
        "operationId": "DriverController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
...
    "/api/v1/drivers/status": {
      "post": {
        "operationId": "DriverController_setStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
    "/api/v1/drivers/location": {
      "post": {
        "operationId": "DriverController_updateLocation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
    "/api/v1/drivers/rides/{id}/accept": {
      "post": {
        "operationId": "DriverController_acceptRide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
    "/api/v1/drivers/rides/{id}/arrive": {
      "post": {
        "operationId": "DriverController_arrivedAtPickup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
    "/api/v1/drivers/rides/{id}/boarded": {
      "post": {
        "operationId": "DriverController_passengerBoarded",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
    "/api/v1/drivers/rides/{id}/start": {
      "post": {
        "operationId": "DriverController_startRide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
    "/api/v1/drivers/rides/{id}/complete": {
      "post": {
        "operationId": "DriverController_completeRide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
    "/api/v1/drivers/rides/{id}/cancel": {
      "post": {
        "operationId": "DriverController_cancelRide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Driver"
        ]
      }
    },
...
    "/api/v1/drivers/wallet/transactions": {
      "get": {
        "operationId": "WalletController_getTransactions",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
...
    "/api/v1/admin/rides/{id}/cancel": {
      "post": {
        "operationId": "AdminController_cancelRide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCancelRideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/admin/captains/{id}/wallet/transactions": {
      "get": {
        "operationId": "AdminCaptainsController_getWalletTransactions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
...
  },
  "info": {
    "title": "RIM WAY API",
    "description": "The RIM WAY Backend API Contract for Passenger and Driver Apps",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EstimateFareDto": {
        "type": "object",
        "properties": {
          "pickupLat": {
            "type": "number"
          },
          "pickupLng": {
            "type": "number"
          },
          "dropoffLat": {
            "type": "number"
          },
          "dropoffLng": {
            "type": "number"
          },
          "serviceTypeId": {
            "type": "string"
          }
        },
        "required": [
          "pickupLat",
          "pickupLng",
          "dropoffLat",
          "dropoffLng",
          "serviceTypeId"
        ]
      },
      "RequestRideDto": {
        "type": "object",
        "properties": {
          "pickupLat": {
            "type": "number"
          },
          "pickupLng": {
            "type": "number"
          },
          "dropoffLat": {
            "type": "number"
          },
          "dropoffLng": {
            "type": "number"
          },
          "serviceTypeId": {
            "type": "string"
          },
          "pickupName": {
            "type": "string"
          },
          "dropoffName": {
            "type": "string"
          }
        },
        "required": [
          "pickupLat",
          "pickupLng",
          "dropoffLat",
          "dropoffLng",
          "serviceTypeId",
          "pickupName",
          "dropoffName"
        ]
      },
      "CancelRideDto": {
        "type": "object",
        "properties": {
          "rideId": {
            "type": "string"
          }
        },
        "required": [
          "rideId"
        ]
      },
      "RegistrationVehicleDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "number",
            "minimum": 1990,
            "maximum": 2035
          },
          "color": {
            "type": "string"
          },
          "plateNumber": {
            "type": "string"
          },
          "serviceTypeId": {
            "type": "string"
          }
        },
        "required": [
          "brand",
          "model",
          "year",
          "color",
          "plateNumber",
          "serviceTypeId"
        ]
      },
      "RegistrationFileDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "fileUrl",
          "fileName",
          "mimeType"
        ]
      },
      "DriverRegistrationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "vehicle": {
            "$ref": "#/components/schemas/RegistrationVehicleDto"
          },
          "documents": {
            "minItems": 4,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationFileDto"
            }
          },
          "vehiclePhotos": {
            "minItems": 4,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationFileDto"
            }
          }
        },
        "required": [
          "name",
          "vehicle",
          "documents",
          "vehiclePhotos"
        ]
      },
      "DriverStatusDto": {
        "type": "object",
        "properties": {
          "isOnline": {
            "type": "boolean"
          }
        },
        "required": [
          "isOnline"
        ]
      },
      "DriverLocationDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
...
      "AdminCancelRideDto": {
        "type": "object",
        "properties": {
          "stateVersion": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "stateVersion"
        ]
      },
...
    }
  }
}